refactor(analysis): split selectors out to their own files.
