+++ lib/containers/isochrone.js
@@ -4,7 +4,8 @@ import {fetchTravelTimeSurface, setIsochroneLonLat} from '../actions/analysis'
 import {removeComponent} from '../actions/map'
 import SinglePointAnalysis from '../components/map/isochrone'
 import {ISOCHRONE_COMPONENT} from '../constants/map'
-import {isochrone, comparisonIsochrone} from '../selectors/isochrone'
+import isochrone from '../selectors/isochrone'
+import comparisonIsochrone from '../selectors/comparison-isochrone'
 
 function mapStateToProps (state, props) {
   const {analysis, scenario, project} = state
