+++ lib/containers/single-point-analysis.js
@@ -5,8 +5,10 @@ import SinglePointAnalysis from '../components/analysis'
 import {fetchTravelTimeSurface, setComparisonScenario, setComparisonInProgress, setIsochroneCutoff, setCurrentIndicator, enterAnalysisMode, exitAnalysisMode, setActiveVariant, clearIsochroneResults, setProfileRequest} from '../actions/analysis'
 import {runAnalysis, load as loadRegionalAnalyses, deleteRegionalAnalysis, setRegionalAnalysisBounds} from '../actions/analysis/regional'
 import {addComponent, removeComponent} from '../actions/map'
-import {accessibility, comparisonAccessibility} from '../selectors/accessibility'
-import {percentileCurves, comparisonPercentileCurves} from '../selectors/percentile-curves'
+import accessibility from '../selectors/accessibility'
+import comparisonAccessibility from '../selectors/comparison-accessibility'
+import percentileCurves from '../selectors/percentile-curves'
+import comparisonPercentileCurves from '../selectors/comparison-percentile-curves'
 
 function mapStateToProps (state, params) {
   const {analysis, mapState, scenario, project} = state
