+++ lib/selectors/accessibility.js
@@ -2,7 +2,7 @@ import {createSelector} from 'reselect'
 
 const POSITION = 9 // ninth (zero-based) item in an array of 19 percentiles, 5th to 95th is the 50th
 
-function computeAccessibility (travelTimeSurface, isochroneCutoff, destinationGrid) {
+export function computeAccessibility (travelTimeSurface, isochroneCutoff, destinationGrid) {
   if (travelTimeSurface == null || destinationGrid == null) return null
 
   let accessibility = 0
@@ -25,16 +25,11 @@ function computeAccessibility (travelTimeSurface, isochroneCutoff, destinationGr
   return accessibility
 }
 
-export const accessibility = createSelector(
+const accessibility = createSelector(
   state => state.analysis.travelTimeSurface,
   state => state.analysis.isochroneCutoff,
   state => state.analysis.destinationGrid,
   computeAccessibility
 )
 
-export const comparisonAccessibility = createSelector(
-  state => state.analysis.comparisonTravelTimeSurface,
-  state => state.analysis.isochroneCutoff,
-  state => state.analysis.destinationGrid,
-  computeAccessibility
-)
+export default accessibility
