+++ lib/selectors/isochrone.js
@@ -6,7 +6,7 @@ const POSITION = 9 // ninth (zero-based) item in an array of 19 percentiles, 5th
 
 // singleValuedSurface, width, height all come from selector defined below and thus must
 // be passed in one argument
-function computeIsochrone (singleValuedSurface, isochroneCutoff) {
+export function computeIsochrone (singleValuedSurface, isochroneCutoff) {
   if (singleValuedSurface == null) return null
 
   const { surface, width, height, west, north, zoom } = singleValuedSurface
@@ -28,7 +28,7 @@ function computeIsochrone (singleValuedSurface, isochroneCutoff) {
  * done separately from isochrone computation because it can be saved when the isochrone cutoff changes
  * when put in a separate selector, memoization will handle this for us.
  */
-function computeSingleValuedSurface (travelTimeSurface) {
+export function computeSingleValuedSurface (travelTimeSurface) {
   if (travelTimeSurface == null) return null
   const singleValuedSurface = new Uint8Array(travelTimeSurface.width * travelTimeSurface.height)
 
@@ -54,19 +54,10 @@ const singleValuedSurface = createSelector(
   computeSingleValuedSurface
 )
 
-export const isochrone = createSelector(
+const isochrone = createSelector(
   singleValuedSurface,
   state => state.analysis.isochroneCutoff,
   computeIsochrone
 )
 
-const comparisonSingleValuedSurface = createSelector(
-  state => state.analysis.comparisonTravelTimeSurface,
-  computeSingleValuedSurface
-)
-
-export const comparisonIsochrone = createSelector(
-  comparisonSingleValuedSurface,
-  state => state.analysis.isochroneCutoff,
-  computeIsochrone
-)
+export default isochrone
