+++ lib/selectors/percentile-curves.js
@@ -3,7 +3,7 @@ import {createSelector} from 'reselect'
 const MAX_TRIP_DURATION = 120
 
 /** percentile curves data is just an array of cumulative accessibility curves for different percentiles. */
-function computePercentileCurves (travelTimeSurface, destinationGrid) {
+export function computePercentileCurves (travelTimeSurface, destinationGrid) {
   if (travelTimeSurface == null || destinationGrid == null) return null
 
   const percentileCurves = []
@@ -39,14 +39,10 @@ function computePercentileCurves (travelTimeSurface, destinationGrid) {
   return percentileCurves
 }
 
-export const percentileCurves = createSelector(
+const percentileCurves = createSelector(
   state => state.analysis.travelTimeSurface,
   state => state.analysis.destinationGrid,
   computePercentileCurves
 )
 
-export const comparisonPercentileCurves = createSelector(
-  state => state.analysis.comparisonTravelTimeSurface,
-  state => state.analysis.destinationGrid,
-  computePercentileCurves
-)
+export default percentileCurves
